fix for the panic in copy_to()
Uint8array's copy_to() panics at the last chunk when the length may
not be the same as the chunk size. This is undesirable. Luckily there
is an "unsafe" method that allows us to copy directly into the
pointer. But it is unsafe. Would be nice if copy_to also took a size
parameter and would have a different logic for panic().